+++ STACpopulator/implementations/CORDEXCMIP6_Ouranos/add_CORDEX6.py
@@ -15,7 +15,6 @@ class CORDEX_STAC_Populator(STACpopulatorBase):
     """Populator that creates STAC objects representing CORDEX data from a THREDDS catalog."""
 
     data_model = Cordex6DataModel
-    item_geometry_model = None  # Unnecessary, but kept for consistency
 
     def create_stac_item(self, item_name: str, item_data: dict[str, Any]) -> dict[str, Any]:
         """Return a STAC item."""