+++ STACpopulator/implementations/HRDPS_CRIM/add_HRDPS.py
@@ -15,7 +15,6 @@ class HRDPSpopulator(STACpopulatorBase):
     """Populator that creates STAC objects representing HRDPS data from a THREDDS catalog."""
 
     data_model = HRDPSDataModel
-    item_geometry_model = None  # Unnecessary, but kept for consistency
 
     def create_stac_item(self, item_name: str, item_data: dict[str, Any]) -> dict[str, Any]:
         """Return a STAC item."""