+++ STACpopulator/implementations/RDPS_CRIM/add_RDPS.py
@@ -15,7 +15,6 @@ class RDPSpopulator(STACpopulatorBase):
     """Populator that creates STAC objects representing RDPS data from a THREDDS catalog."""
 
     data_model = RDPSDataModel
-    item_geometry_model = None  # Unnecessary, but kept for consistency
 
     def create_stac_item(self, item_name: str, item_data: dict[str, Any]) -> dict[str, Any]:
         """Return a STAC item."""