+++ tests/test_standalone_stac_item.py
@@ -13,7 +13,6 @@ import xncml
 from STACpopulator.extensions.cmip6 import CMIP6Helper
 from STACpopulator.extensions.thredds import THREDDSExtension, THREDDSHelper
 from STACpopulator.input import THREDDSLoader
-from STACpopulator.models import GeoJSONPolygon, Geometry
 from STACpopulator.populator_base import STACpopulatorBase
 
 
@@ -48,7 +47,7 @@ def test_standalone_stac_item_thredds_ncml(cur_dir):
         "WMS": f"{thredds_url}/wms/{thredds_path}/{thredds_nc}",
         "NetcdfSubset": f"{thredds_url}/ncss/{thredds_path}/{thredds_nc}/dataset.html",
     }
-    stac_item = CMIP6Helper(attrs, GeoJSONPolygon).stac_item()
+    stac_item = CMIP6Helper(attrs).stac_item()
     thredds_helper = THREDDSHelper(attrs["access_urls"])
     thredds_ext = THREDDSExtension.ext(stac_item)
     thredds_ext.apply(services=thredds_helper.services, links=thredds_helper.links)
@@ -61,8 +60,6 @@ def test_standalone_stac_item_thredds_ncml(cur_dir):
 
 
 class MockedNoSTACUpload(STACpopulatorBase):
-    item_geometry_model = Geometry
-
     def load_config(self):
         # bypass auto-load config
         self._collection_info = {